requirement.txt
astroid==2.3.3
autopep8==1.4.4
colorama==0.4.1
Django==2.2.7
easy-thumbnails==2.6
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mod-wsgi==4.6.8+ap24vc15
mysqlclient==1.4.6
Pillow==6.2.1
pycodestyle==2.5.0
pylint==2.4.4
pytz==2019.3
six==1.13.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
